backend/requirements/base.txt
# Django
Django==5.0.1
django-environ==0.11.2

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Celery
celery==5.3.4
redis==5.0.1
flower==2.0.1

# AWS
boto3==1.34.10
django-storages==1.14.2

# CORS
django-cors-headers==4.3.1

# Utilities
python-dotenv==1.0.0
Pillow==10.1.0
